requirements.txt
# FASTAPI
fastapi
uvicorn

# DATA VALIDATION
pydantic

# LOGGING
loguru

# ENV VARS
python-dotenv

# TESTING
pytest
pytest-asyncio

# DATABASE
sqlmodel
SQLAlchemy
alembic
asyncpg

# SECURITY
PyJWT
bcrypt